SCMLiteApi/app/consumer/requirements.txt
pymongo[srv]
python-dotenv==0.21.1
kafka-python==2.0.2